In pursuit of issue 42:
1. Amend various messages in kbus.c to cast printk arguments appropriately,
   so that (for instance) a %u argument gets an unsigned
2. Amend messages.py to use '=L' instead of 'L' for struct declaration
3. Amend ksock.py to use 'I' instead of 'L' for array declaration
4. In messages.py, correct the "to_bytes" method, add various explanatory
   comments in various places, and do other work as necessary.
5. Amend kbus.c to use the "entire message" structure for the user message,
   so that the padding is correct on 64-bit linux.

Also, in test_kbus.c, amend various tests. Remove a couple of tests that
were very old, and now (a) redundant and (b) hard to make portable.

Fudge check_IOError() for the change in behaviour in 2.6.5 when trying to
write to a read-only file - this now gives a more informative error (for
users), but it is no longer an actual "system" errno, so doesn't set some
of the expected values in the exception.

Note that the handling of REPLIERBIND messages has been fudged for now, until
issue 41 is properly finished...

--HG--
extra : convert_revision : svn%3A53f7ac7a-121e-11de-a074-998af0e899fe/trunk%40410